backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.8.1",
    "django==5.2.3",
    "django-cors-headers==4.7.0",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "pillow>=11.2.1",
    "pyjwt==2.9.0",
    "python-dotenv==1.1.0",
    "sqlparse==0.5.3",
]
